Move episode related fiels to 'episode' directory
