+++ test/auth/auth_test.mocks.dart
@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.3.2 from annotations
+// Mocks generated by Mockito 5.4.0 from annotations
 // in brunstadtv_app/test/auth/auth_test.dart.
 // Do not manually edit this file.
